requirements.txt
flask==2.2.5
requests==2.31.0
certifi==2024.2.2
flask_oauthlib==0.9.6
python_dotenv==0.21.1